+++ CHANGE_LOG.md
@@ -1,5 +1,9 @@
 # CHANGE LOG
 #### All the changes are listed...
 
-### Version: 0.1 [12/08/2023]
-* Initial structure of the library 
+### Version: 0.1 [03/11/2023]
+* Initial structure of the library .
+* publish with (get_market_watch, get_top_gainers, get_heat_map, get_top_gainers, 
+    get_top_losers, get_most_active_contracts, get_most_active_puts_calls, get_bhav_copy, 
+    get_historical_date_wise_data, get_pro_cli_details, get_option_chain, get_put_call_ratio, get_category_wise_oi, 
+    get_category_wise_turnover, get_mcx_icomdex_indices, get_trading_statistics, get_ccl_delivery) functions.